CV_Project/requirements.txt
torch
torchvision
numpy
matplotlib
tqdm
Pillow 